+++ main.py
@@ -14,163 +14,12 @@ from tinydb import TinyDB
 from packages.helper  import Helper
 import os
 
-class KursLoeschenDialog(QtWidgets.QDialog):
-    def __init__(self, kurse):
-        super().__init__()
-        
-
-        self.setWindowTitle("Kurs Löschen")
-        layout = QtWidgets.QVBoxLayout()
-        self.setFixedSize(400, 450)
-        
-        self.kurs_combo = QtWidgets.QComboBox()
-        self.kurs_combo.addItems(kurse)
-        layout.addWidget(self.kurs_combo)
-        
-        button_box = QtWidgets.QDialogButtonBox(QtWidgets.QDialogButtonBox.Ok | QtWidgets.QDialogButtonBox.Cancel)
-        button_box.accepted.connect(self.delete_course)
-        button_box.rejected.connect(self.close_popup)
-        layout.addWidget(button_box)
-        
-        self.setLayout(layout)
-
-    def delete_course(self):
-        current_item = self.kurs_combo.currentText()
-        help.delete_course_item(current_item)
-        self.close()
-        
-    def close_popup(self):
-        self.close()
-
-class ExportScorecardDialog(QtWidgets.QDialog):
-    def __init__(self):
-        super().__init__()
-        self.setWindowTitle("Scorecard Exportieren")
-        layout = QtWidgets.QVBoxLayout()
-
-        self.kurs_combo = QtWidgets.QComboBox()
-        kurse = help.getAllCourseIDs()
-        self.kurs_combo.addItems(kurse)
-        layout.addWidget(QtWidgets.QLabel("Kurs:"))
-        layout.addWidget(self.kurs_combo)
-
-        self.whs_radio = QtWidgets.QRadioButton("WHS")
-        self.ega_radio = QtWidgets.QRadioButton("EGA")
-        self.whs_radio.setChecked(True)
-        layout.addWidget(self.whs_radio)
-        layout.addWidget(self.ega_radio)
-
-        layout.addWidget(QtWidgets.QLabel("Kurs Rating überschreiben:"))
-        self.kurs_rating_eingabe = QtWidgets.QLineEdit()
-        self.kurs_rating_eingabe.setValidator(QtGui.QDoubleValidator())
-        layout.addWidget(self.kurs_rating_eingabe)
-
-        layout.addWidget(QtWidgets.QLabel("Slope Rating überschreiben:"))
-        self.slope_rating_eingabe = QtWidgets.QLineEdit()
-        self.slope_rating_eingabe.setValidator(QtGui.QIntValidator())
-        layout.addWidget(self.slope_rating_eingabe)
-
-        layout.addWidget(QtWidgets.QLabel("Stroke Index überschreiben:"))
-        self.stroke_index_layout = QtWidgets.QGridLayout()
-        self.stroke_index_eingabe = {}
-        self.stroke_index_labels = {}
-        for i in range(1, 19):
-            eingabe = QtWidgets.QLineEdit()
-            eingabe.setValidator(QtGui.QIntValidator())
-            self.stroke_index_eingabe[i] = eingabe
-            label_text = f"Stroke Index Loch {i}:"
-            label = QtWidgets.QLabel(label_text)
-            self.stroke_index_labels[i] = label
-
-            row = (i - 1) % 9
-            col = (i - 1) // 9
-            self.stroke_index_layout.addWidget(label, row, col * 2)
-            self.stroke_index_layout.addWidget(eingabe, row, col * 2 + 1)
-
-            if i > 9:
-                eingabe.hide()
-                label.hide()
-
-        layout.addLayout(self.stroke_index_layout)
-
-        self.file_path_label = QtWidgets.QLabel("Kein Dateipfad ausgewählt")
-        layout.addWidget(self.file_path_label)
-
-        self.file_path_button = QtWidgets.QPushButton("Dateipfad auswählen")
-        self.file_path_button.clicked.connect(self.select_file_path)
-        layout.addWidget(self.file_path_button)
-
-        self.use_last_values_checkbox = QtWidgets.QCheckBox("Die letzten Werte verwenden")
-        layout.addWidget(self.use_last_values_checkbox)
-
-        button_box = QtWidgets.QDialogButtonBox(QtWidgets.QDialogButtonBox.Ok | QtWidgets.QDialogButtonBox.Cancel)
-        button_box.accepted.connect(self.accept)
-        button_box.rejected.connect(self.reject)
-        layout.addWidget(button_box)
-
-        self.setLayout(layout)
-
-        self.kurs_combo.currentIndexChanged.connect(self.update_stroke_index_eingabe)
-
-    def update_stroke_index_eingabe(self):
-        kurs_name = self.kurs_combo.currentText()
-        kurs_info = help.getCourseByID(kurs_name)
-        locher = len(kurs_info["par"])
-        for i in range(1, 19):
-            self.stroke_index_eingabe[i].setVisible(i <= locher)
-            self.stroke_index_labels[i].setVisible(i <= locher)
-
-    def select_file_path(self):
-        options = QtWidgets.QFileDialog.Options()
-        options |= QtWidgets.QFileDialog.DontUseNativeDialog
-        file_path, _ = QtWidgets.QFileDialog.getSaveFileName(self, "Dateipfad auswählen", "", "All Files (*);;Text Files (*.txt)", options=options)
-        if file_path:
-            self.file_path_label.setText(file_path)
-            self.selected_file_path = file_path
-
-    def get_export_daten(self):
-        kurs = self.kurs_combo.currentText()
-
-        kurs_info = help.getCourseByID(kurs)
-
-        is_whs = self.whs_radio.isChecked()
-        kurs_rating = self.kurs_rating_eingabe.text()
-        kurs_rating = float(kurs_rating) if kurs_rating else None
-        slope_rating = self.slope_rating_eingabe.text()
-        slope_rating = int(slope_rating) if slope_rating else None
-        stroke_indices = []
-
-        locher = len(kurs_info["par"])
-
-        for i in range(1, locher + 1):
-            try:
-                stroke_indices.append(int(self.stroke_index_eingabe[i].text()))
-            except ValueError:
-                stroke_indices.append(None)
-
-        if all(index is None for index in stroke_indices):
-            stroke_indices = None
-
-        file_path = getattr(self, 'selected_file_path', None)
-        use_last_values = self.use_last_values_checkbox.isChecked()
-        return kurs, is_whs, file_path, kurs_rating, slope_rating, stroke_indices, use_last_values
-
 class HandicapUI(QtWidgets.QWidget):
     def __init__(self):
         super().__init__()     
         # main component
         self.layout = QtWidgets.QVBoxLayout()
 
-        self.headerButton = HeaderButton(help, self)
-        
-        
-
-        button_header = QtWidgets.QHBoxLayout()
-        self.kurs_löschen_button = QtWidgets.QPushButton("Kurs Löschen")
-        self.export_scorecard_button = QtWidgets.QPushButton("Scorecard Exportieren")
-        button_header.addWidget(self.kurs_löschen_button)
-        button_header.addWidget(self.export_scorecard_button) 
-        self.layout.addLayout(button_header)
 
         self.ega_handicap_label = QtWidgets.QLabel("Aktuelles EGA Handicap: N/A")
         self.ega_handicap_label.setStyleSheet("font-size: 20px; color: white;")
@@ -189,36 +38,26 @@ class HandicapUI(QtWidgets.QWidget):
         self.spiele_tabelle.setHorizontalHeaderLabels(["Datum", "Kurs", "Schläge"])
         self.layout.addWidget(self.spiele_tabelle)
 
-
+        # new component based buttons move to top of init for correct order
+        self.header_button = HeaderButton(help, self)
         self.debug_buttons = DebugButtons(help, self)
+        self.initUI()
 
         self.setLayout(self.layout)
 
-
-        self.kurs_löschen_button.clicked.connect(self.kurs_loeschen_dialog)
-        self.export_scorecard_button.clicked.connect(self.oeffne_export_scorecard_dialog) 
-
-
-        
-        self.initUI()
 
     def initUI(self):
         app.setStyle('Fusion') 
         self.setWindowTitle("Golf Handicap Rechner")
         self.setWindowIcon(QIcon("./mauerIcon.ico"))
         
-        self.layout.addWidget(self.headerButton)
+        self.layout.addWidget(self.header_button)
 
         self.layout.addWidget(self.debug_buttons)
 
 
         self.update()
 
-
-    def kurs_loeschen_dialog(self):
-        kurse = help.getAllCourseIDs()
-        dialog = KursLoeschenDialog(kurse)
-        dialog.exec_()
         
     def oeffne_export_scorecard_dialog(self):
         dialog = ExportScorecardDialog()
